Protect against CLRF injection in unicorn nginx
Our current configuration is vulnerable to CRLF injection whenever we
perform a rewrite, which currently includes the `$uri` directive. `$uri`
attempts to parse and decode the uri, including encoded CRLF characters.
This allows an attacker to insert arbitrary header values.

We fix this by using `$request_uri` instead of `$uri`, which does not
attempt any sort of parsing.